+++ docs/CHANGELOG.md
@@ -31,3 +31,4 @@
 4. createCallBackUrl() added
 5. Ajax test removed 
 6. Test and corect optional Alpaca Services with focuser
+7. Support focuser optional URL's